Make mutant happy again
