+++ app/forms/steps/other_children/names_form.rb
@@ -1,6 +1,6 @@
 module Steps
   module OtherChildren
-    class NamesForm < BaseNamesForm
+    class NamesForm < NamesBaseForm
       private
 
       def record_collection
